requirements.txt
google-generativeai
python-dotenv~=1.1.1
flask~=3.1.1

google~=3.0.0
protobuf~=5.29.5